+++ wiki/release-notes.md
@@ -4,6 +4,38 @@ This page tracks a high-level overview of changes to the version of UCloud
 hosted at cloud.sdu.dk. This page is intended for internal
 use.
 
+## 04-03-26
+
+```
+UCloud/Core @ 2026.2.10
+UCloud/Frontend @ 2026.2.11
+SDU/K8s @ 2026.2.7
+AAU/K8s @ 2026.2.7
+```
+
+## 02-03-26
+
+```
+UCloud/Frontend @ 2026.2.5
+UCloud/Core @ 2026.2.6
+```
+
+## 26-02-26
+
+```
+UCloud/Frontend @ 2026.2.1
+UCloud/Core @ 2026.2.4
+```
+
+## 24-02-26
+
+```
+UCloud/Core @ 2026.2.0
+UCloud/Frontend @ 2026.2.0
+SDU/K8s @ 2026.2.0
+AAU/K8s @ 2026.2.0
+```
+
 ## 16-02-26
 
 ```